+++ maru/consensus/src/main/kotlin/maru/consensus/blockimport/BeaconBlockImporter.kt
@@ -112,6 +112,7 @@ class BlockBuildingBeaconBlockImporter(
   private val prevRandaoProvider: PrevRandaoProvider<ULong>,
   private val shouldBuildNextBlock: (BeaconState, ConsensusRoundIdentifier, ULong) -> Boolean,
   private val feeRecipient: ByteArray,
+  private val targetGasLimit: ULong? = null,
 ) : BeaconBlockImporter {
   private val log: Logger = LogManager.getLogger(this.javaClass)
 
@@ -147,6 +148,8 @@ class BlockBuildingBeaconBlockImporter(
               .inc(),
             prevRandao = beaconBlock.beaconBlockBody.executionPayload.prevRandao,
           ),
+          nextBlockSlotNumber = beaconBlockHeader.number + 1UL,
+          targetGasLimit = targetGasLimit,
         )
       } catch (error: Exception) {
         SafeFuture.failedFuture(error)